requirements.txt
Pillow
PyPDF2
streamlit
python-dotenv
google-generativeai
streamlit
langchain
chromadb
faiss-cup